+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "nondevtaskkiller" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.1]
+
+- Initial release of the extension with Email Refine, Translate, and Jira Refine tools.
+- Added code annotation feature.
+- Refactored the codebase for better readability and maintainability.
+- Added a setting to control the scope of code annotation.
+- Added a command to clear annotations.
+
 ## [Unreleased]
 
 - Initial release